% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBioc-package.R
\docType{package}
\name{testBioc-package}
\alias{testBioc}
\alias{testBioc-package}
\title{testBioc: Test Bioconductor Package Installation}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abichat/testBioc}
  \item Report bugs at \url{https://github.com/abichat/testBioc/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
